site/requirements/local.txt
-r base.txt
django-debug-toolbar==0.9.4
Sphinx==1.2b1
django-discover-runner==1.0
